Restructure integration tests for i18n
